news-analyzer/requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
groq==1.1.2
httpx==0.28.1
praw==7.7.1
prawcore==2.4.0
pytest==8.2.0
python-dotenv==1.0.1
requests==2.31.0
